Rename /containers /app
