+++ src/app/Clash/components/Clash/ClashHeader.js
