+++ src/app/Clash/components/Comments/connectedComments.js
