+++ src/app/Clash/components/CurrentRound/ReadyToAccept.js
