+++ src/app/Clash/components/CurrentRound/Upload.js
