+++ src/app/Clash/components/CurrentRound/connectedUpload.js
