+++ src/app/Clash/components/Round/RoundCompressed.js
