+++ src/app/Home/components/ClashTile.js
